+++ arnold/crud/read/plot/__init__.py
@@ -1,0 +1,4 @@
+from .trend_step_fields import trend_step_fields
+from .trend_turn_around_times import trend_turn_around_times
+from .compare_step_fields import compare_step_fields
+from .trend_nr_recieved_samples import trend_nr_samples_per_month
