+++ arnold/crud/read/plot/compare_step_fields.py
@@ -1,0 +1,42 @@
+from arnold.adapter import ArnoldAdapter
+
+
+def compare_step_fields(
+    adapter: ArnoldAdapter,
+    workflow: str = "TWIST",
+    step_type_x: str = "aliquot_samples_for_enzymatic_fragmentation",
+    step_type_y: str = "capture_and_wash",
+    udf_x: str = "artifact_udfs.amount_needed",
+    udf_y: str = "artifact_udfs.library_size_pre_hyb",
+    # group: str = "sample.application",
+) -> list:
+    pipe = [
+        {
+            "$match": {
+                "$or": [
+                    {"step_type": step_type_x},
+                    {"step_type": step_type_y},
+                ],
+                "workflow": workflow,
+            }
+        },
+        {
+            "$project": {
+                "prep_id": 1,
+                "step_type": 1,
+                udf_x: 1,
+                udf_y: 1,
+            }
+        },
+        {
+            "$group": {
+                "_id": {"prep_id": "$prep_id"},
+                "artifact_udfs": {"$push": "$artifact_udfs"},
+                "step_type": {"$push": "$step_type"},
+            }
+        },
+    ]
+    try:
+        return list(adapter.step_collection.aggregate(pipe))
+    except:
+        return []
